+++ CHANGELOG.md
@@ -19,6 +19,7 @@ All notable, unreleased changes to this project will be documented in this file.
 - Add missing language enum in GraphQL schema - #3854 by @jxltom
 - Fix layout in staff members details page - #3857 by @dominik-zeglen
 - Add draft orders query - #3809 by @michaljelonek
+- Restrict global orders query - #3861 by @pawelzar
 
 
 ## 2.4.0